Compatibility with WTForms 2.3.x
In WTForms 2.3.x the form.errors dictionary is not mutable anymore (it's
autogenerated on access from the fields' errors).
It was never really intended to be mutable but we abused the caching
system, and now we can't do that anymore.
