+++ noggin/form/add_group_member.py
@@ -1,10 +1,11 @@
-from flask_wtf import FlaskForm
 from wtforms import StringField
 from wtforms.validators import DataRequired
 from flask_babel import lazy_gettext as _
 
+from .base import BaseForm
 
-class AddGroupMemberForm(FlaskForm):
+
+class AddGroupMemberForm(BaseForm):
     new_member_username = StringField(
         'New Member Username',
         validators=[DataRequired(message=_('New member username must not be empty'))],
