+++ noggin/form/remove_group_member.py
@@ -1,10 +1,11 @@
 from flask_babel import lazy_gettext as _
-from flask_wtf import FlaskForm
 from wtforms import HiddenField
 from wtforms.validators import DataRequired
 
+from .base import BaseForm
 
-class RemoveGroupMemberForm(FlaskForm):
+
+class RemoveGroupMemberForm(BaseForm):
     username = HiddenField(
         _('Username'),
         validators=[DataRequired(message=_('Username must not be empty'))],
